added modals and splitted the code in different modules
