+++ src/style.css
@@ -74,7 +74,7 @@ header {
   z-index: 3;
   position: fixed;
   left: 0;
-  top: 0;
+  top: 44px !important;
   width: 100%;
   height: 100%;
   overflow: auto;
